+++ com.unity.cinemachine/Editor/PropertyDrawers/LensSettingsPropertyDrawer.cs
@@ -5,12 +5,6 @@ using System.Collections.Generic;
 using UnityEngine.UIElements;
 using UnityEditor.UIElements;
 
-#if CINEMACHINE_HDRP
-    using UnityEngine.Rendering.HighDefinition;
-#elif CINEMACHINE_URP
-    using UnityEngine.Rendering.Universal;
-#endif
-
 namespace Unity.Cinemachine.Editor
 {
     [CustomPropertyDrawer(typeof(LensSettingsHideModeOverridePropertyAttribute))]
@@ -22,7 +16,7 @@ namespace Unity.Cinemachine.Editor
     [CustomPropertyDrawer(typeof(LensSettings))]
     class LensSettingsPropertyDrawer : PropertyDrawer
     {
-        static LensSettings s_LensSettingsDef = new (); // to access name strings
+        static LensSettings s_Def = new (); // to access name strings
 
         static bool IsOrtho(SerializedProperty property) => AccessProperty<bool>(
             typeof(LensSettings), SerializedPropertyHelper.GetPropertyValue(property), "Orthographic");
@@ -30,12 +24,11 @@ namespace Unity.Cinemachine.Editor
         static bool IsPhysical(SerializedProperty property) => AccessProperty<bool>(
             typeof(LensSettings), SerializedPropertyHelper.GetPropertyValue(property), "IsPhysicalCamera");
 
-        static Vector2 SensorSize(SerializedProperty property) => AccessProperty<Vector2>(
-            typeof(LensSettings), SerializedPropertyHelper.GetPropertyValue(property), "SensorSize");
+        static float Aspect(SerializedProperty property) => AccessProperty<float>(
+            typeof(LensSettings), SerializedPropertyHelper.GetPropertyValue(property), "Aspect");
 
-        static bool UseHorizontalFOV(SerializedProperty property) => 
-            InspectorUtility.GetUseHorizontalFOV(AccessProperty<Camera>(
-                typeof(LensSettings), SerializedPropertyHelper.GetPropertyValue(property), "SourceCamera"));
+        static bool UseHorizontalFOV(SerializedProperty property) => AccessProperty<bool>(
+            typeof(LensSettings), SerializedPropertyHelper.GetPropertyValue(property), "UseHorizontalFOV");
 
         static T AccessProperty<T>(Type type, object obj, string memberName)
         {
@@ -56,11 +49,11 @@ namespace Unity.Cinemachine.Editor
                 return default;
         }
 
-        static bool s_PhysicalExapnded;
-
         static List<string> m_PresetOptions;
         static List<string> m_PhysicalPresetOptions;
+        const string k_AddPresetsLabel = "New Preset with these Settings...";
         const string k_EditPresetsLabel = "Edit Presets...";
+        float m_PreviousAspect;
 
         protected bool HideModeOverride { get; set; }
 
@@ -69,16 +62,19 @@ namespace Unity.Cinemachine.Editor
             m_PresetOptions ??= new List<string>();
             m_PresetOptions.Clear();
             var presets = CinemachineLensPresets.InstanceIfExists;
-            for (int i = 0; presets != null && i < presets.m_Presets.Length; ++i)
-                m_PresetOptions.Add(presets.m_Presets[i].m_Name);
+            for (int i = 0; presets != null && i < presets.Presets.Count; ++i)
+                m_PresetOptions.Add(presets.Presets[i].Name);
             m_PresetOptions.Add("");
+            m_PresetOptions.Add(k_AddPresetsLabel);
             m_PresetOptions.Add(k_EditPresetsLabel);
 
+            var physicalPresets = CinemachinePhysicalLensPresets.InstanceIfExists;
             m_PhysicalPresetOptions ??= new List<string>();
             m_PhysicalPresetOptions.Clear();
-            for (int i = 0; presets != null && i < presets.m_PhysicalPresets.Length; ++i)
-                m_PhysicalPresetOptions.Add(presets.m_PhysicalPresets[i].m_Name);
+            for (int i = 0; physicalPresets != null && i < physicalPresets.Presets.Count; ++i)
+                m_PhysicalPresetOptions.Add(physicalPresets.Presets[i].Name);
             m_PhysicalPresetOptions.Add("");
+            m_PhysicalPresetOptions.Add(k_AddPresetsLabel);
             m_PhysicalPresetOptions.Add(k_EditPresetsLabel);
         }
 
@@ -99,7 +95,7 @@ namespace Unity.Cinemachine.Editor
             // Populate the foldout
             var innerFovControl = foldout.AddChild(new FovPropertyControl(property, false) { style = { flexGrow = 1 }});
 
-            var nearClip = property.FindPropertyRelative(() => s_LensSettingsDef.NearClipPlane);
+            var nearClip = property.FindPropertyRelative(() => s_Def.NearClipPlane);
             foldout.AddChild(new PropertyField(nearClip)).RegisterValueChangeCallback((evt) =>
             {
                 if (!IsOrtho(property) && nearClip.floatValue < 0.01f)
@@ -108,52 +104,21 @@ namespace Unity.Cinemachine.Editor
                     property.serializedObject.ApplyModifiedPropertiesWithoutUndo();
                 }
             });
-            foldout.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.FarClipPlane)));
-            foldout.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.Dutch)));
+            foldout.Add(new PropertyField(property.FindPropertyRelative(() => s_Def.FarClipPlane)));
+            foldout.Add(new PropertyField(property.FindPropertyRelative(() => s_Def.Dutch)));
 
-            var physical = foldout.AddChild(new Foldout() { text = "Physical Properties", value = s_PhysicalExapnded });
-            physical.RegisterValueChangedCallback((evt) => 
-            {
-                if (evt.target == physical)
-                    s_PhysicalExapnded = evt.newValue;
-            });
-
-            physical.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.GateFit)));
-
-            var ssProp = property.FindPropertyRelative(() => s_LensSettingsDef.m_SensorSize);
-            var sensorSizeField = physical.AddChild(new PropertyField(ssProp));
-            sensorSizeField.RegisterValueChangeCallback((evt) =>
-            {
-                var v = ssProp.vector2Value;
-                v.x = Mathf.Max(v.x, 0.1f);
-                v.y = Mathf.Max(v.y, 0.1f);
-                ssProp.vector2Value = v;
-                property.serializedObject.ApplyModifiedPropertiesWithoutUndo();
-            });
-
-#if CINEMACHINE_HDRP
-            physical.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.Iso)));
-            physical.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.ShutterSpeed)));
-#endif
-            physical.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.LensShift)));
-            physical.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.FocusDistance)));
-#if CINEMACHINE_HDRP
-            physical.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.Aperture)));
-            physical.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.BladeCount)));
-            physical.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.Curvature)));
-            physical.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.BarrelClipping)));
-            physical.Add(new PropertyField(property.FindPropertyRelative(() => s_LensSettingsDef.Anamorphism)));
-#endif
+            var physical = foldout.AddChild(new PropertyField(property.FindPropertyRelative(() => s_Def.PhysicalProperties)));
 
             SerializedProperty modeOverrideProperty = null;
             VisualElement modeHelp = null;
             if (!HideModeOverride)
             {
-                modeOverrideProperty = property.FindPropertyRelative(() => s_LensSettingsDef.ModeOverride);
+                modeOverrideProperty = property.FindPropertyRelative(() => s_Def.ModeOverride);
                 modeHelp = foldout.AddChild(
                     new HelpBox("Lens Mode Override must be enabled in the CM Brain for Mode Override to take effect", 
                         HelpBoxMessageType.Warning));
-                foldout.Add(new PropertyField(modeOverrideProperty));
+                foldout.AddChild(new PropertyField(modeOverrideProperty)).TrackPropertyValue(
+                    modeOverrideProperty, (p) => InspectorUtility.RepaintGameView());
             }
 
             // GML: This is rather evil.  Is there a better (event-driven) way?
@@ -164,10 +129,16 @@ namespace Unity.Cinemachine.Editor
                 if (property.serializedObject.targetObject == null)
                     return; // target deleted
 
+                // We need to track the aspect ratio because HFOV display is dependent on it
+                var isHorizontal = UseHorizontalFOV(property);
+                var aspect = Aspect(property);
+                bool aspectChanged = isHorizontal && m_PreviousAspect != aspect;
+                m_PreviousAspect = aspect;
+
                 bool isPhysical = IsPhysical(property);
                 physical.SetVisible(isPhysical);
-                outerFovControl.TimedUpdate();
-                innerFovControl.TimedUpdate();
+                outerFovControl.TimedUpdate(aspectChanged);
+                innerFovControl.TimedUpdate(aspectChanged);
 
                 if (!HideModeOverride)
                 {
@@ -177,18 +148,21 @@ namespace Unity.Cinemachine.Editor
                         && modeOverrideProperty.intValue != (int)LensSettings.OverrideModes.None);
                 }
             };
+            // In case presets asset gets deleted or modified externally
+            ux.TrackAnyUserActivity(InitPresetOptions);
 
             return ux;
         }
 
         /// <summary>
-        /// Make the complicated FOV widget which works in 4 modes, with preset popups, 
+        /// Make the complicated FOV widget which works in 4 modes, with preset popups 
         /// and optional weird small label display
         /// </summary>
         class FovPropertyControl : InspectorUtility.LabeledRow
         {
             public readonly Label ShortLabel;
             readonly SerializedProperty m_LensProperty;
+            readonly SerializedProperty m_SensorSizeProperty;
             readonly FloatField m_Control;
             readonly PopupField<string> m_Presets;
 
@@ -205,7 +179,11 @@ namespace Unity.Cinemachine.Editor
             public FovPropertyControl(SerializedProperty property, bool hideLabel) : base(hideLabel ? "" : "(fov)")
             {
                 style.flexDirection = FlexDirection.Row;
+
                 m_LensProperty = property;
+                var physicalProp = property.FindPropertyRelative(() => s_Def.PhysicalProperties);
+                m_SensorSizeProperty = physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.SensorSize);
+
                 m_Control = Contents.AddChild(new FloatField("") { style = {flexBasis = 20, flexGrow = 2, marginLeft = 0}});
                 m_Control.RegisterValueChangedCallback(OnControlValueChanged);
                 Label.SetVisible(!hideLabel);
@@ -230,7 +208,7 @@ namespace Unity.Cinemachine.Editor
                 {
                     case Modes.Ortho:
                     {
-                        var orthoProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.OrthographicSize);
+                        var orthoProp = m_LensProperty.FindPropertyRelative(() => s_Def.OrthographicSize);
                         orthoProp.floatValue = evt.newValue;
                         orthoProp.serializedObject.ApplyModifiedProperties();
                         break;
@@ -238,13 +216,12 @@ namespace Unity.Cinemachine.Editor
                     case Modes.Physical:
                     {
                         // Convert and clamp
-                        var sensorHeight = SensorSize(m_LensProperty).y;
-                        var vfov = Camera.FocalLengthToFieldOfView(Mathf.Max(0.01f, evt.newValue), sensorHeight);
+                        var vfov = FocalLengthToFov(evt.newValue);
                         vfov = Mathf.Clamp(vfov, 1, 179);
-                        m_Control.SetValueWithoutNotify(Camera.FieldOfViewToFocalLength(vfov, sensorHeight));
+                        m_Control.SetValueWithoutNotify(FovToFocalLength(vfov));
 
                         // Push to property
-                        var fovProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.FieldOfView);
+                        var fovProp = m_LensProperty.FindPropertyRelative(() => s_Def.FieldOfView);
                         fovProp.floatValue = vfov;
                         fovProp.serializedObject.ApplyModifiedProperties();
                         break;
@@ -257,12 +234,10 @@ namespace Unity.Cinemachine.Editor
 
                         // Convert from display units
                         if (mode == Modes.HFOV)
-                        {
-                            var sensorSize = SensorSize(m_LensProperty);
-                            newValue = Camera.HorizontalToVerticalFieldOfView(newValue, sensorSize.x / sensorSize.y);
-                        }
+                            newValue = Camera.HorizontalToVerticalFieldOfView(newValue, Aspect(m_LensProperty));
+
                         // Push to property
-                        var fovProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.FieldOfView);
+                        var fovProp = m_LensProperty.FindPropertyRelative(() => s_Def.FieldOfView);
                         fovProp.floatValue = newValue;
                         fovProp.serializedObject.ApplyModifiedProperties();
                         break;
@@ -277,39 +252,48 @@ namespace Unity.Cinemachine.Editor
                 {
                     case Modes.Ortho:
                     {
-                        var orthoProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.OrthographicSize);
+                        var orthoProp = m_LensProperty.FindPropertyRelative(() => s_Def.OrthographicSize);
                         m_Control.SetValueWithoutNotify(orthoProp.floatValue);
                         break;
                     }
                     case Modes.Physical:
                     {
                         // Convert to display FolcalLength units
-                        var fovProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.FieldOfView);
-                        var v = Camera.FieldOfViewToFocalLength(fovProp.floatValue, SensorSize(m_LensProperty).y);
+                        var fovProp = m_LensProperty.FindPropertyRelative(() => s_Def.FieldOfView);
+                        var v = FovToFocalLength(fovProp.floatValue);
                         m_Control.SetValueWithoutNotify(v);
-                        SyncPresetDisplay(m_Control.value, true);
+
+                        // Sync the presets
+                        var presets = CinemachinePhysicalLensPresets.InstanceIfExists;
+                        var index = presets == null ? -1 : presets.GetMatchingPreset(new ()
+                        {
+                            FocalLength = v,
+                            PhysicalProperties = ReadPhysicalSettings()
+                        });
+                        m_Presets.SetValueWithoutNotify(index < 0 ? string.Empty : presets.Presets[index].Name);
                         break;
                     }
                     case Modes.VFOV:
                     case Modes.HFOV:
                     {
                         // Convert to display FOV units
-                        var fovProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.FieldOfView);
+                        var fovProp = m_LensProperty.FindPropertyRelative(() => s_Def.FieldOfView);
                         var v = fovProp.floatValue;
                         if (mode == Modes.HFOV)
-                        {
-                            var sensorSize = SensorSize(m_LensProperty);
-                            v = Camera.VerticalToHorizontalFieldOfView(v, sensorSize.x / sensorSize.y);
-                        }
+                            v = Camera.VerticalToHorizontalFieldOfView(v, Aspect(m_LensProperty));
                         m_Control.SetValueWithoutNotify(v);
-                        SyncPresetDisplay(m_Control.value, false);
+
+                        // Sync the presets
+                        var presets = CinemachineLensPresets.InstanceIfExists;
+                        var index = presets == null ? -1 : presets.GetMatchingPreset(fovProp.floatValue);
+                        m_Presets.SetValueWithoutNotify(index < 0 ? string.Empty : presets.Presets[index].Name);
                         break;
                     }
                 }
-                TimedUpdate();
+                TimedUpdate(false);
             }
 
-            public void TimedUpdate()
+            public void TimedUpdate(bool aspectChanged)
             {
                 switch (GetLensMode())
                 {
@@ -318,7 +302,7 @@ namespace Unity.Cinemachine.Editor
                         var text = "O";
                         if (ShortLabel.text != text)
                         {
-                            var orthoProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.OrthographicSize);
+                            var orthoProp = m_LensProperty.FindPropertyRelative(() => s_Def.OrthographicSize);
                             ShortLabel.text = text;
                             Label.text = orthoProp.displayName;
                             Label.tooltip = m_Control.tooltip = ShortLabel.tooltip = orthoProp.tooltip;
@@ -331,7 +315,7 @@ namespace Unity.Cinemachine.Editor
                         var text = "F";
                         if (ShortLabel.text != text)
                         {
-                            var fovProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.FieldOfView);
+                            var fovProp = m_LensProperty.FindPropertyRelative(() => s_Def.FieldOfView);
                             ShortLabel.text = text;
                             Label.text = ShortLabel.tooltip = "Focal Length";
                             Label.tooltip = m_Control.tooltip = fovProp.tooltip;
@@ -345,13 +329,16 @@ namespace Unity.Cinemachine.Editor
                         var text = "H";
                         if (ShortLabel.text != text)
                         {
-                            var fovProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.FieldOfView);
+                            var fovProp = m_LensProperty.FindPropertyRelative(() => s_Def.FieldOfView);
                             ShortLabel.text = text;
                             Label.text = ShortLabel.tooltip = "Horizontal FOV";
                             Label.tooltip = m_Control.tooltip = fovProp.tooltip;
                             m_Presets.choices = m_PresetOptions;
                             m_Presets.SetVisible(true);
                         }
+                        // No event-driven way to detect this.  Our display depends on aspect
+                        if (aspectChanged)
+                            OnLensPropertyChanged(m_LensProperty);
                         break;
                     }
                     case Modes.VFOV:
@@ -359,7 +346,7 @@ namespace Unity.Cinemachine.Editor
                         var text = "V";
                         if (ShortLabel.text != text)
                         {
-                            var fovProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.FieldOfView);
+                            var fovProp = m_LensProperty.FindPropertyRelative(() => s_Def.FieldOfView);
                             ShortLabel.text = text;
                             Label.text = ShortLabel.tooltip = "Vertical FOV";
                             Label.tooltip = m_Control.tooltip = fovProp.tooltip;
@@ -371,87 +358,127 @@ namespace Unity.Cinemachine.Editor
                 }
             }
 
+            float FocalLengthToFov(float focal) => Camera.FocalLengthToFieldOfView(Mathf.Max(0.01f, focal), m_SensorSizeProperty.vector2Value.y);
+            float FovToFocalLength(float vfov) => Camera.FieldOfViewToFocalLength(vfov, m_SensorSizeProperty.vector2Value.y);
+
             void OnPresetValueChanged(ChangeEvent<string> evt)
             {
-                // Edit the presets assets if desired
-                if (evt.newValue == k_EditPresetsLabel)
-                    Selection.activeObject = CinemachineLensPresets.Instance;
-                else 
+                if (GetLensMode() == Modes.Physical)
                 {
-                    // Apply the preset
-                    var fovProp = m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.FieldOfView);
-                    if (GetLensMode() == Modes.Physical)
+                    // Physical presets
+                    var presets = CinemachinePhysicalLensPresets.Instance;
+                    if (presets != null)
                     {
-                        var index = CinemachineLensPresets.Instance.GetPhysicalPresetIndex(evt.newValue);
-                        if (index >= 0)
+                        // Edit the presets assets if desired
+                        if (evt.newValue == k_EditPresetsLabel)
+                            Selection.activeObject = presets;
+                        else if (evt.newValue == k_AddPresetsLabel)
                         {
-                            var v = CinemachineLensPresets.Instance.m_PhysicalPresets[index];
-                            fovProp.floatValue = Camera.FocalLengthToFieldOfView(
-                                Mathf.Max(0.01f, v.m_FocalLength), SensorSize(m_LensProperty).y);
-#if CINEMACHINE_HDRP
-                            m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.Aperture).floatValue = v.Aperture;
-                            m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.Iso).intValue = v.Iso;
-                            m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.ShutterSpeed).floatValue = v.ShutterSpeed;
-                            m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.BladeCount).intValue = v.BladeCount;
-                            m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.Curvature).vector2Value = v.Curvature;
-                            m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.BarrelClipping).floatValue = v.BarrelClipping;
-                            m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.Anamorphism).floatValue = v.Anamorphism;
-                            m_LensProperty.FindPropertyRelative(() => s_LensSettingsDef.LensShift).vector2Value = v.LensShift;
-#endif
+                            Selection.activeObject = presets;
+                            Undo.RecordObject(presets, "add preset");
+                            presets.Presets.Add(new ()
+                            {
+                                Name = $"{m_Control.value}mm preset {presets.Presets.Count + 1}",
+                                FocalLength = m_Control.value,
+                                PhysicalProperties = ReadPhysicalSettings()
+                            });
+                        }
+                        else 
+                        {
+                            // Apply the preset
+                            var index = presets.GetPresetIndex(evt.newValue);
+                            if (index >= 0)
+                            {
+                                var v = presets.Presets[index];
+                                m_LensProperty.FindPropertyRelative(() => s_Def.FieldOfView).floatValue = FocalLengthToFov(v.FocalLength);
+                                WritePhysicalSettings(v.PhysicalProperties);
+                                m_LensProperty.serializedObject.ApplyModifiedProperties();
+                                return;
+                            }
+                        }
+                    }
+                }
+                else
+                {
+                    // Nonphysical Presets
+                    var presets = CinemachineLensPresets.Instance;
+                    if (presets != null)
+                    {
+                        var fovProp = m_LensProperty.FindPropertyRelative(() => s_Def.FieldOfView);
+
+                        // Edit the presets assets if desired
+                        if (evt.newValue == k_EditPresetsLabel)
+                            Selection.activeObject = presets;
+                        else if (evt.newValue == k_AddPresetsLabel)
+                        {
+                            Selection.activeObject = presets;
+                            Undo.RecordObject(presets, "add preset");
+                            presets.Presets.Add(new ()
+                            {
+                                Name = $"{fovProp.floatValue} preset {presets.Presets.Count + 1}",
+                                VerticalFOV = fovProp.floatValue,
+                            });
+                        }
+                        else 
+                        {
+                            // Apply the preset
+                            var index = presets.GetPresetIndex(evt.newValue);
+                            if (index >= 0)
+                                fovProp.floatValue = presets.Presets[index].VerticalFOV;
                             m_LensProperty.serializedObject.ApplyModifiedProperties();
                         }
                     }
-                    else
-                    {
-                        var index = CinemachineLensPresets.Instance.GetPresetIndex(evt.newValue);
-                        if (index >= 0)
-                        {
-                            fovProp.floatValue = CinemachineLensPresets.Instance.m_Presets[index].m_FieldOfView;
-                            fovProp.serializedObject.ApplyModifiedProperties();
-                        }
-                    }
                 }
+                m_Presets.SetValueWithoutNotify(string.Empty);
             }
 
-            void SyncPresetDisplay(float lensValue, bool physical)
+            LensSettings.PhysicalSettings ReadPhysicalSettings()
             {
-                var v = string.Empty;
-                var presets = CinemachineLensPresets.InstanceIfExists;
-                if (presets != null)
+                var p = m_LensProperty.FindPropertyRelative(() => s_Def.PhysicalProperties);
+                return new ()
                 {
-                    if (physical)
-                    {
-                        var index = presets.GetMatchingPhysicalPreset(lensValue);
-                        if (index >= 0)
-                            v = CinemachineLensPresets.Instance.m_PhysicalPresets[index].m_Name;
-                    }
-                    else
-                    {
-                        var index = presets.GetMatchingPreset(lensValue);
-                        if (index >= 0)
-                            v = CinemachineLensPresets.Instance.m_Presets[index].m_Name;
-                    }
-                }
-                m_Presets.SetValueWithoutNotify(v);
+                    GateFit = (Camera.GateFitMode)p.FindPropertyRelative(() => s_Def.PhysicalProperties.GateFit).intValue,
+                    SensorSize = p.FindPropertyRelative(() => s_Def.PhysicalProperties.SensorSize).vector2Value,
+                    LensShift = p.FindPropertyRelative(() => s_Def.PhysicalProperties.LensShift).vector2Value,
+                    Iso = p.FindPropertyRelative(() => s_Def.PhysicalProperties.Iso).intValue,
+                    ShutterSpeed = p.FindPropertyRelative(() => s_Def.PhysicalProperties.ShutterSpeed).floatValue,
+                    Aperture = p.FindPropertyRelative(() => s_Def.PhysicalProperties.Aperture).floatValue,
+                    BladeCount = p.FindPropertyRelative(() => s_Def.PhysicalProperties.BladeCount).intValue,
+                    Curvature = p.FindPropertyRelative(() => s_Def.PhysicalProperties.Curvature).vector2Value,
+                    BarrelClipping = p.FindPropertyRelative(() => s_Def.PhysicalProperties.BarrelClipping).floatValue,
+                    Anamorphism =p.FindPropertyRelative(() => s_Def.PhysicalProperties.Anamorphism).floatValue
+                };
+            }
+
+            void WritePhysicalSettings(in LensSettings.PhysicalSettings s)
+            {
+                var p = m_LensProperty.FindPropertyRelative(() => s_Def.PhysicalProperties);
+                p.FindPropertyRelative(() => s_Def.PhysicalProperties.GateFit).intValue = (int)s.GateFit;
+                p.FindPropertyRelative(() => s_Def.PhysicalProperties.SensorSize).vector2Value = s.SensorSize;
+                p.FindPropertyRelative(() => s_Def.PhysicalProperties.LensShift).vector2Value = s.LensShift;
+                p.FindPropertyRelative(() => s_Def.PhysicalProperties.Iso).intValue = s.Iso;
+                p.FindPropertyRelative(() => s_Def.PhysicalProperties.ShutterSpeed).floatValue = s.ShutterSpeed;
+                p.FindPropertyRelative(() => s_Def.PhysicalProperties.Aperture).floatValue = s.Aperture;
+                p.FindPropertyRelative(() => s_Def.PhysicalProperties.BladeCount).intValue = s.BladeCount;
+                p.FindPropertyRelative(() => s_Def.PhysicalProperties.Curvature).vector2Value = s.Curvature;
+                p.FindPropertyRelative(() => s_Def.PhysicalProperties.BarrelClipping).floatValue = s.BarrelClipping;
+                p.FindPropertyRelative(() => s_Def.PhysicalProperties.Anamorphism).floatValue = s.Anamorphism;
             }
         }
 
+#if true
         ///===========================================================================
         /// IMGUI IMPLEMENTATION (to be removed) 
         ///===========================================================================
 
-        static readonly GUIContent EditPresetsLabel = new GUIContent("Edit Presets...");
-        static readonly GUIContent HFOVLabel = new GUIContent("Horizontal FOV", "Horizontal Field of View");
-        static readonly GUIContent VFOVLabel = new GUIContent("Vertical FOV", "Vertical Field of View");
-        static readonly GUIContent FocalLengthLabel = new GUIContent("Focal Length", "The length of the lens (in mm)");
-        static readonly GUIContent OrthoSizeLabel = new GUIContent("Ortho Size", "When using an orthographic camera, "
+        static readonly GUIContent EditPresetsLabel = new ("Edit Presets...");
+        static readonly GUIContent HFOVLabel = new ("Horizontal FOV", "Horizontal Field of View");
+        static readonly GUIContent VFOVLabel = new ("Vertical FOV", "Vertical Field of View");
+        static readonly GUIContent FocalLengthLabel = new ("Focal Length", "The length of the lens (in mm)");
+        static readonly GUIContent OrthoSizeLabel = new ("Ortho Size", "When using an orthographic camera, "
             + "this defines the half-height, in world coordinates, of the camera view.");
-        static readonly GUIContent SensorSizeLabel = new GUIContent("Sensor Size", 
-            "Actual size of the image sensor (in mm), used to "
-            + "convert between focal length and field of view.");
-        static readonly GUIContent s_EmptyContent = new GUIContent(" ");
-        static readonly GUIContent PhysicalPropertiesLabel = new GUIContent("Physical Properties", "Physical properties of the lens");
-        static readonly GUIContent AdvancedLabel = new GUIContent("Advanced");
+        static readonly GUIContent s_EmptyContent = new (" ");
+        static readonly GUIContent AdvancedLabel = new ("Advanced");
         static readonly string AdvancedHelpboxMessage = "Lens Mode Override must be enabled in the CM Brain for Mode Override to take effect";
 
         static bool s_AdvancedLensExpanded;
@@ -460,11 +487,8 @@ namespace Unity.Cinemachine.Editor
         {
             public bool IsOrtho;
             public bool IsPhysical;
-            public Vector2 SensorSize;
+            public float Aspect;
             public bool UseHorizontalFOV;
-
-            public GUIContent[] m_PresetOptions;
-            public GUIContent[] m_PhysicalPresetOptions;
         }
         Snapshot m_Snapshot;
 
@@ -473,26 +497,11 @@ namespace Unity.Cinemachine.Editor
 
         void SnapshotCameraShadowValues(SerializedProperty property)
         {
-            if (m_Snapshot.m_PresetOptions == null)
-            {
-                var options = new List<GUIContent>();
-                CinemachineLensPresets presets = CinemachineLensPresets.InstanceIfExists;
-                for (int i = 0; presets != null && i < presets.m_Presets.Length; ++i)
-                    options.Add(new GUIContent(presets.m_Presets[i].m_Name));
-                options.Add(EditPresetsLabel);
-                m_Snapshot.m_PresetOptions = options.ToArray();
-
-                options.Clear();
-                for (int i = 0; presets != null && i < presets.m_PhysicalPresets.Length; ++i)
-                    options.Add(new GUIContent(presets.m_PhysicalPresets[i].m_Name));
-                options.Add(EditPresetsLabel);
-                m_Snapshot.m_PhysicalPresetOptions = options.ToArray();
-            }
             // Assume lens is up-to-date
             m_Snapshot.UseHorizontalFOV = UseHorizontalFOV(property);
             m_Snapshot.IsOrtho = IsOrtho(property);
             m_Snapshot.IsPhysical = IsPhysical(property);
-            m_Snapshot.SensorSize = SensorSize(property);
+            m_Snapshot.Aspect = Aspect(property);
         }
 
         GUIContent GetFOVLabel()
@@ -506,17 +515,13 @@ namespace Unity.Cinemachine.Editor
         {
             if (m_Snapshot.IsOrtho)
                 EditorGUI.PropertyField(
-                    rect, property.FindPropertyRelative(() => s_LensSettingsDef.OrthographicSize), label);
+                    rect, property.FindPropertyRelative(() => s_Def.OrthographicSize), label);
             else if (m_Snapshot.IsPhysical)
                 DrawFocalLengthControl(rect, property, label);
             else
             {
-                var FOVProperty = property.FindPropertyRelative(() => s_LensSettingsDef.FieldOfView);
-                float aspect = m_Snapshot.SensorSize.x / m_Snapshot.SensorSize.y;
-
-                float dropdownWidth = (rect.width - EditorGUIUtility.labelWidth) / 3;
-                rect.width -= dropdownWidth + hSpace;
-
+                var FOVProperty = property.FindPropertyRelative(() => s_Def.FieldOfView);
+                float aspect = m_Snapshot.Aspect;
                 float f = FOVProperty.floatValue;
                 if (m_Snapshot.UseHorizontalFOV)
                     f = Camera.VerticalToHorizontalFieldOfView(f, aspect);
@@ -527,100 +532,22 @@ namespace Unity.Cinemachine.Editor
                 if (!Mathf.Approximately(FOVProperty.floatValue, f))
                     FOVProperty.floatValue = Mathf.Clamp(f, 1, 179);
                 EditorGUI.EndProperty();
-                rect.x += rect.width + hSpace; rect.width = dropdownWidth;
-
-                CinemachineLensPresets presets = CinemachineLensPresets.InstanceIfExists;
-                int preset = (presets == null) ? -1 : presets.GetMatchingPreset(FOVProperty.floatValue);
-
-                var oldLabelWidth = EditorGUIUtility.labelWidth;
-                EditorGUIUtility.labelWidth = 1;
-                int selection = EditorGUI.Popup(rect, s_EmptyContent, preset, m_Snapshot.m_PresetOptions);
-                EditorGUIUtility.labelWidth = oldLabelWidth;
-                if (selection == m_Snapshot.m_PresetOptions.Length-1 && CinemachineLensPresets.Instance != null)
-                    Selection.activeObject = presets = CinemachineLensPresets.Instance;
-                else if (selection >= 0 && selection < m_Snapshot.m_PresetOptions.Length-1)
-                {
-                    var vfov = presets.m_Presets[selection].m_FieldOfView;
-                    FOVProperty.floatValue = vfov;
-                    property.serializedObject.ApplyModifiedProperties();
-                }
             }
         }
 
         void DrawFocalLengthControl(Rect rect, SerializedProperty property, GUIContent label)
         {
-            const float hSpace = 2;
-            var FOVProperty = property.FindPropertyRelative(() => s_LensSettingsDef.FieldOfView);
-            float dropdownWidth = (rect.width - EditorGUIUtility.labelWidth) / 4;
-            rect.width -= dropdownWidth + hSpace;
+            var FOVProperty = property.FindPropertyRelative(() => s_Def.FieldOfView);
+            var physicalProp = property.FindPropertyRelative(() => s_Def.PhysicalProperties);
+            var sensorSizeProp = physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.SensorSize);
 
-            float f = Camera.FieldOfViewToFocalLength(FOVProperty.floatValue, m_Snapshot.SensorSize.y);
+            float f = Camera.FieldOfViewToFocalLength(FOVProperty.floatValue, sensorSizeProp.vector2Value.y);
             EditorGUI.BeginProperty(rect, label, FOVProperty);
             f = EditorGUI.FloatField(rect, label, f);
-            f = Camera.FocalLengthToFieldOfView(Mathf.Max(0.01f, f), m_Snapshot.SensorSize.y);
+            f = Camera.FocalLengthToFieldOfView(Mathf.Max(0.01f, f), sensorSizeProp.vector2Value.y);
             if (!Mathf.Approximately(FOVProperty.floatValue, f))
                 FOVProperty.floatValue = Mathf.Clamp(f, 1, 179);
             EditorGUI.EndProperty();
-
-            rect.x += rect.width + hSpace; rect.width = dropdownWidth;
-
-#if CINEMACHINE_HDRP
-            CinemachineLensPresets presets = CinemachineLensPresets.InstanceIfExists;
-            int preset = -1;
-            if (presets != null)
-            {
-                var focalLength = Camera.FieldOfViewToFocalLength(FOVProperty.floatValue, m_Snapshot.SensorSize.y);
-                var aperture = property.FindPropertyRelative(() => s_LensSettingsDef.Aperture).floatValue;
-                var iso = property.FindPropertyRelative(() => s_LensSettingsDef.Iso).intValue;
-                var shutterSpeed = property.FindPropertyRelative(() => s_LensSettingsDef.ShutterSpeed).floatValue;
-                var bladeCount = property.FindPropertyRelative(() => s_LensSettingsDef.BladeCount).intValue;
-                var curvature = property.FindPropertyRelative(() => s_LensSettingsDef.Curvature).vector2Value;
-                var barrelClipping = property.FindPropertyRelative(() => s_LensSettingsDef.BarrelClipping).floatValue;
-                var anamprphism = property.FindPropertyRelative(() => s_LensSettingsDef.Anamorphism).floatValue;
-                var lensShift = property.FindPropertyRelative(() => s_LensSettingsDef.LensShift).vector2Value;
-
-                preset = presets.GetMatchingPhysicalPreset(
-                    focalLength, iso, shutterSpeed, aperture, bladeCount,
-                    curvature, barrelClipping, anamprphism, lensShift);
-            }
-
-            var oldLabelWidth = EditorGUIUtility.labelWidth;
-            EditorGUIUtility.labelWidth = 1;
-            int selection = EditorGUI.Popup(rect, s_EmptyContent, preset, m_Snapshot.m_PhysicalPresetOptions);
-            EditorGUIUtility.labelWidth = oldLabelWidth;
-            if (selection == m_Snapshot.m_PhysicalPresetOptions.Length-1 && CinemachineLensPresets.Instance != null)
-                Selection.activeObject = presets = CinemachineLensPresets.Instance;
-            else if (selection >= 0 && selection < m_Snapshot.m_PhysicalPresetOptions.Length-1)
-            {
-                var v = presets.m_PhysicalPresets[selection];
-                FOVProperty.floatValue = Camera.FocalLengthToFieldOfView(Mathf.Max(0.01f, v.m_FocalLength), m_Snapshot.SensorSize.y);
-                property.FindPropertyRelative(() => s_LensSettingsDef.Aperture).floatValue = v.Aperture;
-                property.FindPropertyRelative(() => s_LensSettingsDef.Iso).intValue = v.Iso;
-                property.FindPropertyRelative(() => s_LensSettingsDef.ShutterSpeed).floatValue = v.ShutterSpeed;
-                property.FindPropertyRelative(() => s_LensSettingsDef.BladeCount).intValue = v.BladeCount;
-                property.FindPropertyRelative(() => s_LensSettingsDef.Curvature).vector2Value = v.Curvature;
-                property.FindPropertyRelative(() => s_LensSettingsDef.BarrelClipping).floatValue = v.BarrelClipping;
-                property.FindPropertyRelative(() => s_LensSettingsDef.Anamorphism).floatValue = v.Anamorphism;
-                property.FindPropertyRelative(() => s_LensSettingsDef.LensShift).vector2Value = v.LensShift;
-                property.serializedObject.ApplyModifiedProperties();
-            }
-#else
-            CinemachineLensPresets presets = CinemachineLensPresets.InstanceIfExists;
-            int preset = (presets == null) ? -1 : presets.GetMatchingPhysicalPreset(
-                Camera.FieldOfViewToFocalLength(FOVProperty.floatValue, m_Snapshot.SensorSize.y));
-            var oldLabelWidth = EditorGUIUtility.labelWidth;
-            EditorGUIUtility.labelWidth = 1;
-            int selection = EditorGUI.Popup(rect, s_EmptyContent, preset, m_Snapshot.m_PhysicalPresetOptions);
-            EditorGUIUtility.labelWidth = oldLabelWidth;
-            if (selection == m_Snapshot.m_PhysicalPresetOptions.Length-1 && CinemachineLensPresets.Instance != null)
-                Selection.activeObject = presets = CinemachineLensPresets.Instance;
-            else if (selection >= 0 && selection < m_Snapshot.m_PhysicalPresetOptions.Length-1)
-            {
-                FOVProperty.floatValue = Camera.FocalLengthToFieldOfView(
-                    Mathf.Max(0.01f, presets.m_PhysicalPresets[selection].m_FocalLength), m_Snapshot.SensorSize.y);
-                property.serializedObject.ApplyModifiedProperties();
-            }
-#endif
         }
 
         public override void OnGUI(Rect rect, SerializedProperty property, GUIContent label)
@@ -656,7 +583,7 @@ namespace Unity.Cinemachine.Editor
                 DrawFOVControl(rect, property, fovLabel);
 
                 rect.y += rect.height + vSpace;
-                var nearClip = property.FindPropertyRelative(() => s_LensSettingsDef.NearClipPlane);
+                var nearClip = property.FindPropertyRelative(() => s_Def.NearClipPlane);
                 EditorGUI.PropertyField(rect, nearClip);
                 if (!m_Snapshot.IsOrtho && nearClip.floatValue < 0.01f)
                 {
@@ -664,50 +591,39 @@ namespace Unity.Cinemachine.Editor
                     property.serializedObject.ApplyModifiedPropertiesWithoutUndo();
                 }
                 rect.y += rect.height + vSpace;
-                EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.FarClipPlane));
+                EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_Def.FarClipPlane));
 
                 rect.y += rect.height + vSpace;
-                EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.Dutch));
+                EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_Def.Dutch));
 
                 if (m_Snapshot.IsPhysical)
                 {
+                    var physicalProp = property.FindPropertyRelative(() => s_Def.PhysicalProperties);
                     rect.y += rect.height + vSpace;
-                    s_PhysicalExapnded = EditorGUI.Foldout(rect, s_PhysicalExapnded, PhysicalPropertiesLabel, true);
-
-                    if (s_PhysicalExapnded)
+                    physicalProp.isExpanded = EditorGUI.Foldout(rect, physicalProp.isExpanded, physicalProp.displayName, true);
+                    if (physicalProp.isExpanded)
                     {
                         ++EditorGUI.indentLevel;
-
                         rect.y += rect.height + vSpace;
-                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.GateFit));
-
+                        EditorGUI.PropertyField(rect, physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.GateFit));
                         rect.y += rect.height + vSpace;
-                        var ssProp = property.FindPropertyRelative(() => s_LensSettingsDef.m_SensorSize);
-                        EditorGUI.BeginProperty(rect, SensorSizeLabel, ssProp);
-                        var v = EditorGUI.Vector2Field(rect, SensorSizeLabel, ssProp.vector2Value);
-                        v.x = Mathf.Max(v.x, 0.1f);
-                        v.y = Mathf.Max(v.y, 0.1f);
-                        ssProp.vector2Value = v;
-                        EditorGUI.EndProperty();
-
+                        EditorGUI.PropertyField(rect, physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.SensorSize));
                         rect.y += rect.height + vSpace;
-                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.LensShift));
-#if CINEMACHINE_HDRP
+                        EditorGUI.PropertyField(rect, physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.LensShift));
                         rect.y += rect.height + vSpace;
-                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.Aperture));
+                        EditorGUI.PropertyField(rect, physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.Iso));
                         rect.y += rect.height + vSpace;
-                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.Iso));
+                        EditorGUI.PropertyField(rect, physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.ShutterSpeed));
                         rect.y += rect.height + vSpace;
-                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.ShutterSpeed));
+                        EditorGUI.PropertyField(rect, physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.Aperture));
                         rect.y += rect.height + vSpace;
-                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.BladeCount));
+                        EditorGUI.PropertyField(rect, physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.BladeCount));
                         rect.y += rect.height + vSpace;
-                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.Curvature));
+                        EditorGUI.PropertyField(rect, physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.Curvature));
                         rect.y += rect.height + vSpace;
-                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.BarrelClipping));
+                        EditorGUI.PropertyField(rect, physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.BarrelClipping));
                         rect.y += rect.height + vSpace;
-                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.Anamorphism));
-#endif
+                        EditorGUI.PropertyField(rect, physicalProp.FindPropertyRelative(() => s_Def.PhysicalProperties.Anamorphism));
                         --EditorGUI.indentLevel;
                     }
                 }
@@ -723,7 +639,7 @@ namespace Unity.Cinemachine.Editor
                         EditorGUI.HelpBox(r, AdvancedHelpboxMessage, MessageType.Info);
 
                         rect.y += r.height + vSpace;
-                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_LensSettingsDef.ModeOverride));
+                        EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => s_Def.ModeOverride));
                         --EditorGUI.indentLevel;
                     }
                 }
@@ -744,14 +660,9 @@ namespace Unity.Cinemachine.Editor
             if (m_Snapshot.IsPhysical)
             {
                 numLines += 1;
-                if (s_PhysicalExapnded)
-                {
-#if CINEMACHINE_HDRP
+                var physicalProp = property.FindPropertyRelative(() => s_Def.PhysicalProperties);
+                if (physicalProp.isExpanded)
                     numLines += 10;
-#else
-                    numLines += 3;
-#endif
-                }
             }
             if (!HideModeOverride)
             {
@@ -762,5 +673,6 @@ namespace Unity.Cinemachine.Editor
             }
             return lineHeight + numLines * (lineHeight + vSpace);
         }
+#endif
     }
 }