+++ com.unity.cinemachine/Editor/Utility/CinemachineSceneTools.cs
@@ -178,7 +178,7 @@ namespace Unity.Cinemachine.Editor
                 if (lens.IsPhysicalCamera)
                 {
                     DrawLabel(labelPos, "Focal Length (" + 
-                        Camera.FieldOfViewToFocalLength(lens.FieldOfView, lens.SensorSize.y).ToString("F1") + ")");
+                        Camera.FieldOfViewToFocalLength(lens.FieldOfView, lens.PhysicalProperties.SensorSize.y).ToString("F1") + ")");
                 }
                 else if (orthographic)
                 {