+++ com.unity.cinemachine/Runtime/PostProcessing/CinemachineAutoFocus.cs
@@ -160,7 +160,7 @@ namespace Unity.Cinemachine
                     focusDistance = extra.CurrentFocusDistance + Damper.Damp(
                         focusDistance - extra.CurrentFocusDistance, Damping, deltaTime);
                 extra.CurrentFocusDistance = focusDistance;
-                state.Lens.FocusDistance = focusDistance;
+                state.Lens.PhysicalProperties.FocusDistance = focusDistance;
             }
         }