+++ com.unity.cinemachine/CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Several settings moved to Transition Settings section of inspector.
 - TrackedObjectOffset renamed to TargetOffset.
 - Improved layout of PositionComposer and RotationComposer inspectors.
+- LensSettings was refactored to improve handling of SensorSize and other physical properties.
+- Full physical camera support for builtin pipeline.
+- LensPresets and PhysicalLensPresets are now separate assets.
 
 
 ## [3.0.0-pre.4] - 2023-02-09